Feature: add client from snapwp.config.ts
