+++ packages/apollo/src/query-engine/index.ts
@@ -10,9 +10,13 @@ import {
 	type QueryHookOptions,
 	type TypedDocumentNode,
 	type OperationVariables,
+	ApolloError,
+	type ServerError,
+	type ServerParseError,
 } from '@apollo/client';
 import type { QueryClientAdapter } from '@snapwp/types';
 import { getGraphqlUrl } from '@snapwp/core/config';
+import { Logger } from '@snapwp/core';
 
 /**
  *
@@ -101,20 +105,37 @@ export class ApolloQueryClientAdapter
 		query: DocumentNode | TypedDocumentNode< TData >;
 		options?: TQueryOptions;
 	} ): Promise< TData > {
-		const { data } = await this.client.query< TData >( {
-			query,
-			...options,
-		} );
+		try {
+			const { data } = await this.client.query< TData >( {
+				query,
+				...options,
+			} );
 
-		this.client.writeQuery< TData >( {
-			query,
-			// @ts-ignore
-			data,
-			id: key.join( ':' ),
-			variables: options?.variables,
-		} );
+			this.client.writeQuery< TData >( {
+				query,
+				// @ts-ignore
+				data,
+				id: key.join( ':' ),
+				variables: options?.variables,
+			} );
 
-		return data;
+			return data;
+		} catch ( error ) {
+			if ( error instanceof ApolloError ) {
+				logApolloErrors( error );
+
+				// If there are networkError throw the error with proper message.
+				if ( error.networkError ) {
+					// Throw the error with proper message.
+					throw new Error(
+						getNetworkErrorMessage( error.networkError )
+					);
+				}
+			}
+
+			// If error is not an instance of ApolloError, throw the error again.
+			throw error;
+		}
 	}
 
 	/**
@@ -144,3 +165,62 @@ export class ApolloQueryClientAdapter
 		} as QueryHookOptions< TData > ) as TData;
 	}
 }
+
+/**
+ * Logs the Apollo errors.
+ *
+ * @param error - The Apollo error.
+ */
+const logApolloErrors = ( error: ApolloError ): void => {
+	// If there are graphQLErrors log them.
+	error.graphQLErrors.forEach( ( graphQLError ) => {
+		Logger.error( graphQLError.message );
+	} );
+
+	// If there are clientErrors log them.
+	error.clientErrors.forEach( ( clientError ) => {
+		Logger.error( clientError.message );
+	} );
+
+	// If there are protocolErrors log them.
+	error.protocolErrors.forEach( ( protocolError ) => {
+		Logger.error( protocolError.message );
+	} );
+};
+
+/**
+ * Returns the network error message.
+ *
+ * @param networkError - The network error.
+ *
+ * @return The network error message.
+ */
+const getNetworkErrorMessage = (
+	networkError: Error | ServerParseError | ServerError
+): string => {
+	let statusCode: number | undefined;
+	let errorMessage: string | undefined;
+	// If networkError is ServerError, get the status code and message.
+	if ( networkError.name === 'ServerError' ) {
+		const serverError = networkError as ServerError;
+		statusCode = serverError.statusCode;
+		if ( typeof serverError.result === 'string' ) {
+			errorMessage = serverError.result;
+		} else {
+			errorMessage = serverError.result[ 'message' ];
+		}
+	} else if (
+		// If networkError is ServerParseError, get the status code and message.
+		networkError.name === 'ServerParseError'
+	) {
+		const serverParseError = networkError as ServerParseError;
+
+		statusCode = serverParseError.statusCode;
+		errorMessage = serverParseError.message;
+	} else {
+		// If networkError is not ServerError or ServerParseError, get the message.
+		errorMessage = networkError.message;
+	}
+
+	return `Network error ${ errorMessage } (Status: ${ statusCode })`;
+};
