+++ packages/next/package.json
@@ -44,8 +44,7 @@
 		"react": "^19.0.0",
 		"react-dom": "^19.0.0",
 		"@snapwp/core": "*",
-		"@snapwp/query": "*",
-		"@snapwp/apollo": "*"
+		"@snapwp/query": "*"
 	},
 	"dependencies": {
 		"html-react-parser": "^5.1.12",
