+++ packages/next/tsconfig.json
@@ -14,7 +14,6 @@
 	"references": [
 		{ "path": "../core" },
 		{ "path": "../query" },
-		{ "path": "../types" },
-		{ "path": "../apollo" }
+		{ "path": "../types" }
 	]
 }
