+++ packages/query/src/query-engine/index.ts
@@ -1,59 +1,37 @@
-import { getGraphqlUrl, getConfig } from '@snapwp/core/config';
 import {
 	GetCurrentTemplateDocument,
 	GetGeneralSettingsDocument,
 	GetGlobalStylesDocument,
 } from '@graphqlTypes/graphql';
-import {
-	ApolloClient,
-	ApolloError,
-	InMemoryCache,
-	type NormalizedCacheObject,
-	type ServerError,
-	type ServerParseError,
-} from '@apollo/client';
 import parseTemplate from '@/utils/parse-template';
 import parseGlobalStyles from '@/utils/parse-global-styles';
-import type { BlockData } from '@snapwp/types';
 import {
-	Logger,
 	type GlobalHeadProps,
+	type ScriptModuleProps,
 	type EnqueuedScriptProps,
 	type StyleSheetProps,
-	type ScriptModuleProps,
 } from '@snapwp/core';
 import parseGeneralSettings from '@/utils/parse-general-settings';
+import { getConfig } from '@snapwp/core/config';
+import { QueryAdapterRegistry } from '@/query-adapter-registry';
+import type { BlockData } from '@snapwp/types';
 
 /**
  * Singleton class to handle GraphQL queries using Apollo.
  */
 export class QueryEngine {
 	private static instance: QueryEngine | null = null;
-	private static graphqlEndpoint: string;
-	private static homeUrl: string;
-	private static apolloClient: ApolloClient< NormalizedCacheObject >;
-
-	private static isClientInitialized = false;
 
 	/**
-	 * Initializer.
+	 * Private constructor to prevent instantiation.
 	 */
-	public static initialize(): void {
-		QueryEngine.graphqlEndpoint = getGraphqlUrl();
-
-		const { wpHomeUrl } = getConfig();
-		QueryEngine.homeUrl = wpHomeUrl;
-
-		QueryEngine.apolloClient = new ApolloClient( {
-			uri: QueryEngine.graphqlEndpoint,
-			cache: new InMemoryCache(),
-		} );
-	}
+	// eslint-disable-next-line no-useless-constructor,no-empty-function -- Constructor is private to prevent instantiation.
+	private constructor() {}
 
 	/**
 	 * Returns the singleton instance of QueryEngine.
-	 * @throws Throws error if instance is not initialized with config.
 	 *
+	 * @throws Throws error if instance is not initialized with config.
 	 * @return The QueryEngine instance.
 	 */
 	public static getInstance(): QueryEngine {
@@ -65,38 +43,16 @@ export class QueryEngine {
 
 	/**
 	 * Fetches global styles.
-	 *
 	 * @return The template data fetched for the uri.
 	 */
-	static getGlobalStyles = async (): Promise< GlobalHeadProps > => {
-		if ( ! QueryEngine.isClientInitialized ) {
-			QueryEngine.initialize();
-		}
+	getGlobalStyles = async (): Promise< GlobalHeadProps > => {
+		const data = await QueryAdapterRegistry.adapter.fetchQuery( {
+			key: [ 'globalStyles' ],
+			query: GetGlobalStylesDocument,
+		} );
 
-		try {
-			const data = await QueryEngine.apolloClient.query( {
-				query: GetGlobalStylesDocument,
-				fetchPolicy: 'network-only', // @todo figure out a caching strategy, instead of always fetching from network
-				errorPolicy: 'all',
-			} );
-
-			return parseGlobalStyles( data );
-		} catch ( error ) {
-			if ( error instanceof ApolloError ) {
-				logApolloErrors( error );
-
-				// If there are networkError throw the error with proper message.
-				if ( error.networkError ) {
-					// Throw the error with proper message.
-					throw new Error(
-						getNetworkErrorMessage( error.networkError )
-					);
-				}
-			}
-
-			// If error is not an instance of ApolloError, throw the error again.
-			throw error;
-		}
+		// @ts-ignore
+		return parseGlobalStyles( data );
 	};
 
 	/**
@@ -104,7 +60,7 @@ export class QueryEngine {
 	 *
 	 * @return General settings data.
 	 */
-	static getGeneralSettings = async (): Promise<
+	getGeneralSettings = async (): Promise<
 		| {
 				generalSettings: {
 					siteIcon: {
@@ -121,43 +77,21 @@ export class QueryEngine {
 		  }
 		| undefined
 	> => {
-		if ( ! QueryEngine.isClientInitialized ) {
-			QueryEngine.initialize();
-		}
+		const data = await QueryAdapterRegistry.adapter.fetchQuery( {
+			key: [ 'generalSettings' ],
+			query: GetGeneralSettingsDocument,
+		} );
 
-		try {
-			const data = await QueryEngine.apolloClient.query( {
-				query: GetGeneralSettingsDocument,
-				fetchPolicy: 'network-only', // @todo figure out a caching strategy, instead of always fetching from network
-				errorPolicy: 'all',
-			} );
-
-			return parseGeneralSettings( data );
-		} catch ( error ) {
-			if ( error instanceof ApolloError ) {
-				logApolloErrors( error );
-
-				// If there are networkError throw the error with proper message.
-				if ( error.networkError ) {
-					// Throw the error with proper message.
-					throw new Error(
-						getNetworkErrorMessage( error.networkError )
-					);
-				}
-			}
-
-			// If error is not an instance of ApolloError, throw the error again.
-			throw error;
-		}
+		// @ts-ignore
+		return parseGeneralSettings( data );
 	};
 
 	/**
 	 * Fetches blocks, scripts and styles for the given uri.
 	 * @param uri - The URL of the seed node.
-	 *
 	 * @return The template data fetched for the uri.
 	 */
-	static getTemplateData = async (
+	getTemplateData = async (
 		uri: string
 	): Promise< {
 		stylesheets: StyleSheetProps[] | undefined;
@@ -166,94 +100,18 @@ export class QueryEngine {
 		scriptModules: ScriptModuleProps[] | undefined;
 		bodyClasses: string[] | undefined;
 	} > => {
-		if ( ! QueryEngine.isClientInitialized ) {
-			QueryEngine.initialize();
-		}
 		const variables = { uri };
 
-		try {
-			const data = await QueryEngine.apolloClient.query( {
-				query: GetCurrentTemplateDocument,
+		const data = await QueryAdapterRegistry.adapter.fetchQuery( {
+			key: [ 'templateData', uri ],
+			query: GetCurrentTemplateDocument,
+			options: {
 				variables,
-				fetchPolicy: 'network-only', // @todo figure out a caching strategy, instead of always fetching from network
-				errorPolicy: 'all',
-			} );
+			},
+		} );
 
-			return parseTemplate( data, QueryEngine.homeUrl, uri );
-		} catch ( error ) {
-			if ( error instanceof ApolloError ) {
-				logApolloErrors( error );
-
-				// If there are networkError throw the error with proper message.
-				if ( error.networkError ) {
-					// Throw the error with proper message.
-					throw new Error(
-						getNetworkErrorMessage( error.networkError )
-					);
-				}
-			}
-
-			// If error is not an instance of ApolloError, throw the error again.
-			throw error;
-		}
+		const { wpHomeUrl } = getConfig();
+		// @ts-ignore
+		return parseTemplate( data, wpHomeUrl, uri );
 	};
 }
-
-/**
- * Logs the Apollo errors.
- *
- * @param error - The Apollo error.
- */
-const logApolloErrors = ( error: ApolloError ): void => {
-	// If there are graphQLErrors log them.
-	error.graphQLErrors.forEach( ( graphQLError ) => {
-		Logger.error( graphQLError.message );
-	} );
-
-	// If there are clientErrors log them.
-	error.clientErrors.forEach( ( clientError ) => {
-		Logger.error( clientError.message );
-	} );
-
-	// If there are protocolErrors log them.
-	error.protocolErrors.forEach( ( protocolError ) => {
-		Logger.error( protocolError.message );
-	} );
-};
-
-/**
- * Returns the network error message.
- *
- * @param networkError - The network error.
- *
- * @return The network error message.
- */
-const getNetworkErrorMessage = (
-	networkError: Error | ServerParseError | ServerError
-): string => {
-	let statusCode: number | undefined;
-	let errorMessage: string | undefined;
-	// If networkError is ServerError, get the status code and message.
-	if ( networkError.name === 'ServerError' ) {
-		const serverError = networkError as ServerError;
-		statusCode = serverError.statusCode;
-		if ( typeof serverError.result === 'string' ) {
-			errorMessage = serverError.result;
-		} else {
-			errorMessage = serverError.result[ 'message' ];
-		}
-	} else if (
-		// If networkError is ServerParseError, get the status code and message.
-		networkError.name === 'ServerParseError'
-	) {
-		const serverParseError = networkError as ServerParseError;
-
-		statusCode = serverParseError.statusCode;
-		errorMessage = serverParseError.message;
-	} else {
-		// If networkError is not ServerError or ServerParseError, get the message.
-		errorMessage = networkError.message;
-	}
-
-	return `Network error ${ errorMessage } (Status: ${ statusCode })`;
-};
